requirements.txt
streamlit>=1.18
google-generativeai>=0.8.5
transformers>=4.44.0
torch
sentencepiece
easyocr
opencv-python
pyzbar
pandas
pillow
rapidfuzz
pytesseract
python-dotenv
